@@ -1,26 +1,25 @@
 <?php
 
-use Faker\Generator as Faker;
+namespace LaravelEnso\Core\Database\Factories;
+
+use Illuminate\Database\Eloquent\Factories\Factory;
 use LaravelEnso\Core\Models\User;
 use LaravelEnso\Core\Models\UserGroup;
 use LaravelEnso\People\Models\Person;
 use LaravelEnso\Roles\Models\Role;
 
-$factory->define(User::class, function (Faker $faker, $attributes = null) {
-    if (isset($attributes['person_id']) && isset($attributes['email'])) {
-        $personId = $attributes['person_id'];
-        $email = $attributes['email'];
-    } else {
-        $person = factory(Person::class)->create();
-        $personId = $person->id;
-        $email = $person->email;
-    }
+class UserFactory extends Factory
+{
+    protected $model = User::class;
 
-    return [
-        'person_id' => $personId,
-        'group_id' => fn () => factory(UserGroup::class)->create()->id,
-        'email' => $email,
-        'role_id' => fn () => factory(Role::class)->create()->id,
-        'is_active' => $faker->boolean,
-    ];
-});
+    public function definition()
+    {
+        return [
+            'person_id' => Person::factory(),
+            'group_id' => UserGroup::factory(),
+            'email' => fn ($attributes) => Person::find($attributes['person_id'])->email,
+            'role_id' => Role::factory(),
+            'is_active' => $this->faker->boolean,
+        ];
+    }
+}
