@@ -1,17 +1,22 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use LaravelEnso\Core\Http\Controllers\Auth\ForgotPasswordController;
+use LaravelEnso\Core\Http\Controllers\Auth\LoginController;
+use LaravelEnso\Core\Http\Controllers\Auth\ResetPasswordController;
 
-Route::namespace('Auth')
-    ->middleware('api')
+Route::middleware('api')
     ->group(function () {
         Route::middleware('guest')->group(function () {
-            Route::post('login', 'LoginController@login')->name('login');
-            Route::post('password/email', 'ForgotPasswordController@sendResetLinkEmail')->name('password.email');
-            Route::post('password/reset', 'ResetPasswordController@attemptReset')->name('password.reset');
+            Route::post('login', [LoginController::class, 'login'])
+                ->name('login');
+            Route::post('password/email', [ForgotPasswordController::class, 'sendResetLinkEmail'])
+                ->name('password.email');
+            Route::post('password/reset', [ResetPasswordController::class, 'attemptReset'])
+                ->name('password.reset');
         });
 
         Route::middleware('auth')->group(function () {
-            Route::post('logout', 'LoginController@logout')->name('logout');
+            Route::post('logout', [LoginController::class, 'logout'])->name('logout');
         });
     });
