@@ -1,11 +1,12 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use LaravelEnso\Core\Http\Controllers\Preferences\Reset;
+use LaravelEnso\Core\Http\Controllers\Preferences\Store;
 
-Route::namespace('Preferences')
-    ->prefix('preferences')
+Route::prefix('preferences')
     ->as('preferences.')
     ->group(function () {
-        Route::patch('store/{route?}', 'Store')->name('store');
-        Route::post('reset/{route?}', 'Reset')->name('reset');
+        Route::patch('store/{route?}', Store::class)->name('store');
+        Route::post('reset/{route?}', Reset::class)->name('reset');
     });
